WIP: isolate declarative charting logic
